Modularized the Onvo ruby SDK
